+++ crates/swc_ecma_codegen/tests/fixture/typescript/class_prop/output.js
@@ -5,6 +5,6 @@ class MyClass extends Base {
     #prop4?: string = "test";
     static readonly prop5!: string;
     readonly #prop6 = "asdf";
-    public override readonly prop7 = 5;
+    public abstract override readonly prop7 = 5;
     override readonly #prop8 = 5;
 }