+++ crates/swc_ecma_transforms_base/tests/ts-resolver/stc/1/output.ts
@@ -1,11 +1,11 @@
 interface Mixin1__2 {
     mixinMethod__0(): void;
 }
-class AbstractBase__2 {
-    abstractBaseMethod(): void;
+abstract class AbstractBase__2 {
+    abstract abstractBaseMethod(): void;
 }
 function Mixin2__2<TBase__4 extends abstract new(...args: any[]) => any>(baseClass__4: TBase__4) {
-    class MixinClass__4 extends baseClass__4 implements Mixin1__2 {
+    abstract class MixinClass__4 extends baseClass__4 implements Mixin1__2 {
         mixinMethod(): void {}
         static staticMixinMethod(): void {}
     }